tech(REC-182) : move out some fucntion from app.js
